changed structure, used react vite stuff
