+++ frontend/src/styles/register.css
