+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "MERN-Project-Gym-App",
+  "name": "mern-project-gym-app",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
